Add support for showing effects for windows which have already been deleted. Add a fade-out effect.
svn path=/branches/work/kwin_composite/; revision=623872

+++ client.cpp
@@ -108,7 +108,6 @@ Client::Client( Workspace *ws )
 
     shade_mode = ShadeNone;
     active = false;
-    deleting = false;
     keep_above = false;
     keep_below = false;
     motif_noborder = false;
@@ -175,9 +174,11 @@ void Client::deleteClient( Client* c, allowed_t )
  */
 void Client::releaseWindow( bool on_shutdown )
     {
-    assert( !deleting );
-    deleting = true;
-    finishCompositing();
+    assert( !deleting());
+    delete_refcount = 1;
+    if( effects )
+        effects->windowClosed( this );
+    finishCompositing( false ); // don't discard pixmap
     workspace()->discardUsedWindowRules( this, true ); // remove ForceTemporarily rules
     StackingUpdatesBlocker blocker( workspace());
     if (moveResizeMode)
@@ -192,7 +193,9 @@ void Client::releaseWindow( bool on_shutdown )
     if( !on_shutdown )
         workspace()->clientHidden( this );
     XUnmapWindow( display(), frameId()); // destroying decoration would cause ugly visual effect
-    destroyDecoration();
+//    destroyDecoration();
+    delete decoration;
+    decoration = NULL;
     cleanGrouping();
     if( !on_shutdown )
         {
@@ -200,9 +203,11 @@ void Client::releaseWindow( bool on_shutdown )
         // only when the window is being unmapped, not when closing down KWin
         // (NETWM sections 5.5,5.7)
         info->setDesktop( 0 );
-        desk = 0;
+        // desk = 0; - do not reset internal state, it may still be used by effects
         info->setState( 0, info->state()); // reset all state flags
         }
+    else
+        workspace()->addDeleted( this, Allowed );
     XDeleteProperty( display(), client, atoms->kde_net_wm_user_creation_time);
     XDeleteProperty( display(), client, atoms->net_frame_extents );
     XDeleteProperty( display(), client, atoms->kde_net_wm_frame_strut );
@@ -220,22 +225,25 @@ void Client::releaseWindow( bool on_shutdown )
         // may do map+unmap before we initially map the window by calling rawShow() from manage().
         XUnmapWindow( display(), client ); 
         }
+    cleanUp();
     client = None;
     XDestroyWindow( display(), wrapper );
     wrapper = None;
     XDestroyWindow( display(), frameId());
 //    frame = None;
     --block_geometry_updates; // don't use GeometryUpdatesBlocker, it would now set the geometry
-    deleteClient( this, Allowed );
+    unrefWindow(); // will delete if recount is == 0
     }
 
 // like releaseWindow(), but this one is called when the window has been already destroyed
 // (e.g. the application closed it)
 void Client::destroyClient()
     {
-    assert( !deleting );
-    deleting = true;
-    finishCompositing();
+    assert( !deleting());
+    delete_refcount = 1;
+    if( effects )
+        effects->windowClosed( this );
+    finishCompositing( false ); // don't discard pixmap
     workspace()->discardUsedWindowRules( this, true ); // remove ForceTemporarily rules
     StackingUpdatesBlocker blocker( workspace());
     if (moveResizeMode)
@@ -247,15 +255,50 @@ void Client::destroyClient()
     setModal( false );
     hidden = true; // so that it's not considered visible anymore
     workspace()->clientHidden( this );
-    destroyDecoration();
+//    destroyDecoration();
+    delete decoration;
+    decoration = NULL;
     cleanGrouping();
     workspace()->removeClient( this, Allowed );
+    cleanUp();
     client = None; // invalidate
     XDestroyWindow( display(), wrapper );
     wrapper = None;
     XDestroyWindow( display(), frameId());
 //    frame = None;
     --block_geometry_updates; // don't use GeometryUpdatesBlocker, it would now set the geometry
+    unrefWindow(); // will delete if recount is == 0
+    }
+
+// All clean-up code shared between releaseWindow() and destroyClient().
+// Clean up everything that should not affect a deleted Client until it's
+// really deleted by unrefWindow(). I.e. stop watching events, turn off timers
+// and so on.
+void Client::cleanUp()
+    {
+    if( Extensions::shapeAvailable())
+        XShapeSelectInput( display(), client, NoEventMask );
+    XSelectInput( display(), client, NoEventMask );
+    delete autoRaiseTimer;
+    autoRaiseTimer = NULL;
+    delete shadeHoverTimer;
+    shadeHoverTimer = NULL;
+    delete ping_timer;
+    ping_timer = NULL;
+    delete process_killer;
+    process_killer = NULL;
+    delete demandAttentionKNotifyTimer;
+    demandAttentionKNotifyTimer = NULL;
+    }
+
+void Client::unrefWindow()
+    {
+    if( --delete_refcount > 0 )
+        return;
+    workspace()->removeDeleted( this );
+    workspace()->addDamage( geometry());
+    destroyDecoration( true ); // only now gravitate etc., otherwise window pixmap would be wrong
+    discardWindowPixmap();
     deleteClient( this, Allowed );
     }
 
@@ -300,9 +343,9 @@ void Client::updateDecoration( bool check_workspace_pos, bool force )
     addDamageFull();
     }
 
-void Client::destroyDecoration()
+void Client::destroyDecoration( bool force )
     {
-    if( decoration != NULL )
+    if( decoration != NULL || force )
         {
         delete decoration;
         decoration = NULL;
@@ -317,7 +360,7 @@ void Client::destroyDecoration()
         workarea_diff_y = save_workarea_diff_y;
         if( compositing() )
             discardWindowPixmap();
-        if( scene != NULL )
+        if( scene != NULL && !deleting())
             scene->windowGeometryShapeChanged( this );
         addDamageFull();
         }
@@ -852,7 +895,7 @@ void Client::toggleShade()
 
 void Client::updateVisibility()
     {
-    if( deleting )
+    if( deleting())
         return;
     bool show = true;
     if( hidden )
@@ -910,7 +953,7 @@ void Client::updateVisibility()
 void Client::setMappingState(int s)
     {
     assert( client != None );
-    assert( !deleting || s == WithdrawnState );
+    assert( !deleting() || s == WithdrawnState );
     if( mapping_state == s )
         return;
     bool was_unmanaged = ( mapping_state == WithdrawnState );
